refactor  getUserProfile and updateUserProfile api implementation
